+++ bodhi/client/__init__.py
@@ -28,6 +28,7 @@ import functools
 
 import click
 import six
+import munch
 
 from bodhi.client import bindings
 from fedora.client import AuthError
@@ -95,6 +96,7 @@ save_edit_options = [
     staging_option,
     url_option]
 
+
 # Basic options for pagination of query result
 pagination_options = [
     click.option('--rows', default=None,
@@ -103,6 +105,33 @@ pagination_options = [
     click.option('--page', default=None,
                  type=click.IntRange(1, clamp=False),
                  help='Go to page number')]
+
+
+# Common releases options
+release_options = [
+    click.option('--username'),
+    click.option('--password', prompt=True, hide_input=True),
+    click.option('--name', help='Release name (eg: F20)'),
+    click.option('--long-name', help='Long release name (eg: "Fedora 20")'),
+    click.option('--id-prefix', help='Release prefix (eg: FEDORA)'),
+    click.option('--version', help='Release version number (eg: 20)'),
+    click.option('--branch', help='Git branch name (eg: f20)'),
+    click.option('--dist-tag', help='Koji dist tag (eg: f20)'),
+    click.option('--stable-tag', help='Koji stable tag (eg: f20-updates)'),
+    click.option('--testing-tag',
+                 help='Koji testing tag (eg: f20-updates-testing)'),
+    click.option('--candidate-tag',
+                 help='Koji candidate tag (eg: f20-updates-candidate)'),
+    click.option('--pending-stable-tag',
+                 help='Koji pending tag (eg: f20-updates-pending)'),
+    click.option('--pending-testing-tag',
+                 help='Koji pending testing tag (eg: f20-updates-testing-testing)'),
+    click.option('--override-tag', help='Koji override tag (eg: f20-override)'),
+    click.option('--state', type=click.Choice(['disabled', 'pending', 'current',
+                                               'archived']),
+                 help='The state of the release'),
+    staging_option,
+    url_option]
 
 
 def add_options(options):
@@ -785,6 +814,139 @@ def print_resp(resp, client, verbose=False):
         click.echo('Caveats:')
         for caveat in resp.caveats:
             click.echo(caveat.description)
+
+
+@cli.group()
+def releases():
+    # Docs that show in the --help
+    """Interact with releases."""
+    # Developer Docs
+    """Manage the releases."""
+    pass  # pragma: no cover
+
+
+@releases.command(name='create')
+@handle_errors
+@add_options(release_options)
+def create_release(username, password, url, **kwargs):
+    """Create a release."""
+    client = bindings.BodhiClient(base_url=url, username=username, password=password,
+                                  staging=kwargs['staging'])
+    kwargs['csrf_token'] = client.csrf()
+
+    save(client, **kwargs)
+
+
+@releases.command(name='edit')
+@handle_errors
+@add_options(release_options)
+@click.option('--new-name', help='New release name (eg: F20)')
+def edit_release(username, password, url, **kwargs):
+    """Edit an existing release."""
+    client = bindings.BodhiClient(base_url=url, username=username, password=password,
+                                  staging=kwargs['staging'])
+    csrf = client.csrf()
+
+    edited = kwargs.pop('name')
+
+    if edited is None:
+        print("ERROR: Please specify the name of the release to edit")
+        return
+
+    res = client.send_request('releases/%s' % edited, verb='GET', auth=True)
+
+    data = munch.unmunchify(res)
+
+    if 'errors' in data:
+        print_errors(data)
+
+    data['edited'] = edited
+    data['csrf_token'] = csrf
+
+    new_name = kwargs.pop('new_name')
+
+    if new_name is not None:
+        data['name'] = new_name
+
+    for k, v in kwargs.items():
+        if v is not None:
+            data[k] = v
+
+    save(client, **data)
+
+
+@releases.command(name='info')
+@handle_errors
+@click.argument('name')
+@url_option
+@staging_option
+def info_release(name, url, **kwargs):
+    """Retrieve and print info about a named release."""
+    client = bindings.BodhiClient(base_url=url, staging=kwargs['staging'])
+
+    res = client.send_request('releases/%s' % name, verb='GET', auth=False)
+
+    if 'errors' in res:
+        print_errors(res)
+
+    else:
+        print('Release:')
+        print_release(res)
+
+
+def save(client, **kwargs):
+    """
+    Save a new or edited release.
+
+    Args:
+        client (bodhi.client.bindings.BodhiClient): The Bodhi client to use for the request.
+        kwargs (dict): The parameters to send with the request.
+    """
+    res = client.send_request('releases/', verb='POST', auth=True,
+                              data=kwargs)
+
+    if 'errors' in res:
+        print_errors(res)
+
+    else:
+        print("Saved release:")
+        print_release(res)
+
+
+def print_release(release):
+    """
+    Print a given release to the terminal.
+
+    Args:
+        release (munch.Munch): The release to be printed.
+    """
+    print("  Name:                %s" % release['name'])
+    print("  Long Name:           %s" % release['long_name'])
+    print("  Version:             %s" % release['version'])
+    print("  Branch:              %s" % release['branch'])
+    print("  ID Prefix:           %s" % release['id_prefix'])
+    print("  Dist Tag:            %s" % release['dist_tag'])
+    print("  Stable Tag:          %s" % release['stable_tag'])
+    print("  Testing Tag:         %s" % release['testing_tag'])
+    print("  Candidate Tag:       %s" % release['candidate_tag'])
+    print("  Pending Signing Tag: %s" % release['pending_signing_tag'])
+    print("  Pending Testing Tag: %s" % release['pending_testing_tag'])
+    print("  Pending Stable Tag:  %s" % release['pending_stable_tag'])
+    print("  Override Tag:        %s" % release['override_tag'])
+    print("  State:               %s" % release['state'])
+
+
+def print_errors(data):
+    """
+    Print errors to the terminal and exit with code 1.
+
+    Args:
+        errors (munch.Munch): The errors to be formatted and printed.
+    """
+    for error in data['errors']:
+        print("ERROR: %s" % error['description'])
+
+    sys.exit(1)
 
 
 if __name__ == '__main__':